get bit

+++ bit_manipulation/2-get_bit.c
@@ -1,7 +1,7 @@
 #include "main.h"
 
 /**
-* git_bit - return the value of a bit at a given index
+* get_bit - return the value of a bit at a given index
 * @n: number
 * @index: the index starting from 0 of the bit
 *
@@ -10,7 +10,7 @@
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int bit;
+	unsigned long int bit;
 
 	bit = (n >> index) & 1;
